refactor: separate types & throw error
