+++ src/types/index.ts
@@ -1,0 +1,11 @@
+export interface StateType {
+  count: number;
+  next: string;
+  previous: null;
+  results: IItem[] | null;
+}
+
+interface IItem {
+  name: string;
+  url: string;
+}
